% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_diagnostics}
\alias{cluster_diagnostics}
\alias{cluster_stability}
\title{Classification Diagnostics}
\usage{
cluster_diagnostics(results)

cluster_stability(results)
}
\arguments{
\item{results}{A `saqr_clustering` object.}
}
\value{
A `multilpa_diagnostics` object (from \pkg{latents}).
}
\description{
Classification Diagnostics
}
