% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_best_model}
\alias{get_best_model}
\alias{cluster_best}
\title{Get Best Model Name}
\usage{
get_best_model(results, criterion = "bic", what = c("name", "fit"))

cluster_best(results, criterion = "bic", what = c("name", "fit"))
}
\arguments{
\item{results}{A `saqr_clustering` object.}

\item{criterion}{Selection criterion.}

\item{what}{What to return: `"name"` or `"fit"`.}
}
\value{
A character string (model name) or a `multilpa` fit.
}
\description{
Get Best Model Name
}
