% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatenets.R
\name{print.grouped_network_analysis}
\alias{print.grouped_network_analysis}
\title{Print method for grouped_network_analysis objects}
\usage{
\method{print}{grouped_network_analysis}(x, ...)
}
\arguments{
\item{x}{A grouped_network_analysis object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for grouped_network_analysis objects
}
