% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{print.html_table}
\alias{print.html_table}
\title{Print method for html tables}
\usage{
\method{print}{html_table}(x, ...)
}
\arguments{
\item{x}{An html_table object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for html tables
}
