% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{print.latex_table}
\alias{print.latex_table}
\title{Print method for latex tables}
\usage{
\method{print}{latex_table}(x, ...)
}
\arguments{
\item{x}{A latex_table object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for latex tables
}
