% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{print.markdown_table}
\alias{print.markdown_table}
\title{Print method for markdown tables}
\usage{
\method{print}{markdown_table}(x, ...)
}
\arguments{
\item{x}{A markdown_table object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for markdown tables
}
