% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mosaic.R
\name{print.mosaic_analysis}
\alias{print.mosaic_analysis}
\title{Print method for mosaic_analysis objects}
\usage{
\method{print}{mosaic_analysis}(x, ...)
}
\arguments{
\item{x}{A mosaic_analysis object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for mosaic_analysis objects
}
