% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatenets.R
\name{print.network_analysis}
\alias{print.network_analysis}
\title{Print method for network_analysis objects}
\usage{
\method{print}{network_analysis}(x, ...)
}
\arguments{
\item{x}{A network_analysis object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for network_analysis objects
}
