% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{print.saqr_clustering}
\alias{print.saqr_clustering}
\title{Print a Clustering Result}
\usage{
\method{print}{saqr_clustering}(x, ...)
}
\arguments{
\item{x}{A `saqr_clustering` object from [clustering()].}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print a Clustering Result
}
