% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.R
\name{print.saqr_result}
\alias{print.saqr_result}
\title{Print method for saqr_result}
\usage{
\method{print}{saqr_result}(x, ...)
}
\arguments{
\item{x}{A saqr_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for saqr_result
}
