% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.R
\name{set_claude_key}
\alias{set_claude_key}
\title{Set Anthropic (Claude) API Key}
\usage{
set_claude_key(key)
}
\arguments{
\item{key}{Your Anthropic API key}
}
\description{
Convenience alias for `set_api_key(key, "anthropic")`.
}
\examples{
\dontrun{
set_claude_key("sk-ant-...")
}

}
