% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.R
\name{set_gemini_key}
\alias{set_gemini_key}
\title{Set Google Gemini API Key}
\usage{
set_gemini_key(key)
}
\arguments{
\item{key}{Your Gemini API key}
}
\description{
Convenience alias for `set_api_key(key, "gemini")`.
}
\examples{
\dontrun{
set_gemini_key("AIza...")
}

}
