% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.R
\name{set_openai_key}
\alias{set_openai_key}
\title{Set OpenAI API Key}
\usage{
set_openai_key(key)
}
\arguments{
\item{key}{Your OpenAI API key}
}
\description{
Convenience alias for `set_api_key(key, "openai")`.
}
\examples{
\dontrun{
set_openai_key("sk-...")
}

}
