% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mosaic.R
\name{summary.mosaic_analysis}
\alias{summary.mosaic_analysis}
\title{Summary method for mosaic_analysis objects}
\usage{
\method{summary}{mosaic_analysis}(object, ...)
}
\arguments{
\item{object}{A mosaic_analysis object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for mosaic_analysis objects
}
