% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatenets.R
\name{summary.network_analysis}
\alias{summary.network_analysis}
\title{Summary method for network_analysis objects}
\usage{
\method{summary}{network_analysis}(object, ...)
}
\arguments{
\item{object}{A network_analysis object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for network_analysis objects
}
