% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_aliases}
\alias{clustering_aliases}
\alias{cluster_fit}
\alias{cluster_models}
\alias{cluster_best}
\alias{cluster_compare}
\alias{cluster_compare_table}
\alias{cluster_assignments}
\alias{cluster_view}
\alias{cluster_plot_model}
\alias{cluster_plot_best}
\alias{cluster_stability}
\alias{cluster_report}
\title{Optional \code{cluster_*} aliases}
\description{
These aliases provide a consistent naming family without replacing or
deprecating the original function names.
}
\details{
\tabular{ll}{
\code{cluster_fit()} \tab \code{clustering()}\cr
\code{cluster_models()} \tab \code{list_models()}\cr
\code{cluster_best()} \tab \code{get_best_model()}\cr
\code{cluster_compare()} \tab \code{compare_models()}\cr
\code{cluster_compare_table()} \tab \code{model_comparison_table()}\cr
\code{cluster_assignments()} \tab \code{get_cluster_assignments()}\cr
\code{cluster_view()} \tab \code{view_results()}\cr
\code{cluster_plot_model()} \tab \code{plot_model()}\cr
\code{cluster_plot_best()} \tab \code{plot_best_model()}\cr
\code{cluster_stability()} \tab \code{assess_cluster_stability()}\cr
\code{cluster_report()} \tab \code{generate_cluster_report()}
}
}
