% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_best_model}
\alias{get_best_model}
\title{Get Best Model}
\usage{
get_best_model(results, criterion = "bic")
}
\arguments{
\item{results}{Object from run_full_moe_analysis}

\item{criterion}{Selection criterion: "bic", "aic", or "icl"}
}
\value{
Name of best model
}
\description{
Get Best Model
}
