% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clustering.R
\name{plot.saqr_clustering}
\alias{plot.saqr_clustering}
\title{Plot Method for Clustering Results}
\usage{
\method{plot}{saqr_clustering}(x, type = "clusters", ...)
}
\arguments{
\item{x}{A `saqr_clustering` object from [clustering()].}

\item{type}{Plot types and/or groups; see [plot_clustering()].}

\item{...}{Passed to [plot_clustering()].}
}
\value{
As [plot_clustering()]: a ggplot object, or a `saqr_plots` list.
}
\description{
Plot Method for Clustering Results
}
